@@ -1,3 +1,3 @@
-import { DbMongo } from "utils/dbMongo";
+import { DbMongo } from "src/utils/dbMongo";
 
 export const API: DbMongo = new DbMongo();
