@@ -1,1 +1,1 @@
-export const apiUrl: string = process.env.REACT_APP_API_URL || "";
+export const apiUrl: string = process.env.REACT_APP_API_URL || '';
